refactor: Project 조회 성공 스키마 위치 변경
