+++ src/main/java/com/waggle/global/response/swagger/ProjectSuccessResponse.java
@@ -1,6 +1,7 @@
-package com.waggle.global.response;
+package com.waggle.global.response.swagger;
 
 import com.waggle.domain.project.entity.Project;
+import com.waggle.global.response.SuccessResponse;
 
 public class ProjectSuccessResponse extends SuccessResponse<Project> {
     public ProjectSuccessResponse(int code, String message, Project payload) {
